backend/requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.4
duckdb==1.2.2
fastapi==0.116.1
lxml==6.0.0
openpyxl==3.1.5
pandas==2.3.1
pyarrow==21.0.0
pydantic-settings==2.11.0
python-multipart==0.0.20
requests==2.32.4
tenacity==9.1.2
uvicorn==0.35.0
xlrd==2.0.1
